+++ rust/crates/spring-native/src/teams.rs
@@ -27,7 +27,7 @@ pub struct PlayerInfo {
     pub team_id: i32,
     pub ally_team_id: i32,
     pub ping_time: f32,
-    pub cpu_usage: u32,
+    pub cpu_usage: f32,
     pub country: String,
     pub rank: i32,
     pub has_skirmish_ais_in_team: bool,
